feat: make S3 bucket URL configurable
Removes hardcoded S3 links that were scattered throughout the site and
replaces them with an option in our JSON config files. I tried to match
the app's design patterns by relying on React context and higher order
components when possible. Non-React utility and constants files import
the local config files using `getDefaultConfig`.

I named the variable `fileServerDomain` because it'll be the domain name
of whatever service we end up using to serve the files. We're still deciding
if a proxy will be in front of the new S3 bucket and it's likely we
won't use raw S3 URLs anymore.

See GHE/HMDA-Operations/hmda-devops/issues/5119