test: add smoke tests script w/ subset of e2e tests
Tags a dozenish e2e tests with a @smoke tag that is triggered with a new
`test-smoke` npm script. Many of the e2e test files perform a series of checks
against multiple years. This results in a long run time. The smoke tests
run against just 2024 and only one spec per page/feature of the frontend.

Closes https://github.com/cfpb/hmda-frontend/issues/2627